+++ R/tadaa_pairwise_games_howell.R
@@ -50,7 +50,7 @@ tadaa_pairwise_gh <- function(data, response, group1, group2 = NULL, print = "df
     return(tests)
   } else {
     output <- pixiedust::dust(tests)
-    output <- pixiedust::sprinkle(output, cols = "adj.p.value", fn = quote(pval_string(value)))
+    output <- pixiedust::sprinkle(output, cols = "adj.p.value", fn = quote(tadaatoolbox::pval_string(value)))
     output <- pixiedust::sprinkle_colnames(output, adj.p.value = "p (adj.)")
     #output <- pixiedust::sprinkle_table(output, cols = 1, caption = "", part = "head")
   }